used dependency injection, renamed evaluable tree, began pre evaluation
